requirements.txt
asgiref==3.4.1
cassandra-driver==3.25.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
geomet==0.2.1.post1
idna==3.3
pycparser==2.20
PyJWT==1.7.1
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
